requirements_windows_stable.txt
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.28
scikit-learn==1.3.2
tensorflow-cpu==2.15.0
matplotlib==3.6.3
seaborn==0.12.2
plotly==5.17.0
streamlit==1.28.1
ta==0.10.2
scipy==1.11.4
xgboost==2.0.3
lightgbm==4.1.0


